+++ server/utils/multer.js
@@ -1,0 +1,28 @@
+const multer = require('multer');
+const path = require("path");
+
+// Check File Type
+function checkFileType(file, cb) {
+    // Allowed ext
+    const filetypes = /jpeg|jpg|png|gif/;
+    // Check ext
+    const extname = filetypes.test(path.extname(file.originalname).toLowerCase());
+    // Check mime
+    const mimetype = filetypes.test(file.mimetype);
+
+    if (mimetype && extname) {
+        return cb(null, true);
+    } else {
+        cb('Error: Images Only!');
+    }
+}
+
+// Multer config
+const upload = multer({
+    storage: multer.diskStorage({}),
+    fileFilter: function (req, file, cb) {
+        checkFileType(file, cb);
+    },
+});
+
+module.exports = upload
